+++ app/services/activities/import_from_csv.rb
@@ -129,9 +129,6 @@ module Activities
         @activity.reporting_organisation = @organisation
 
         beis = Organisation.find_by(service_owner: true)
-        @activity.funding_organisation_name = beis.name
-        @activity.funding_organisation_reference = beis.iati_reference
-        @activity.funding_organisation_type = beis.organisation_type
         @activity.accountable_organisation_name = beis.name
         @activity.accountable_organisation_reference = beis.iati_reference
         @activity.accountable_organisation_type = beis.organisation_type
@@ -288,7 +285,6 @@ module Activities
         attributes[:recipient_region] ||= inferred_region
         attributes[:call_present] = (@row["Call open date"] && @row["Call close date"]).present?
         attributes[:sector_category] = get_sector_category(attributes[:sector])
-        attributes[:status] = infer_status(attributes)
         attributes[:form_state] = "complete"
 
         attributes
@@ -379,6 +375,7 @@ module Activities
         parent = Activity.by_roda_identifier(roda_id)
 
         raise I18n.t("importer.errors.activity.parent_not_found") if parent.nil?
+        raise I18n.t("importer.errors.activity.invalid_parent") unless parent.form_steps_completed?
 
         parent.id
       end
@@ -516,10 +513,6 @@ module Activities
 
           country_to_region_mapping.find { |pair| pair["country"] == @row["Recipient Country"] }["region"]
         end
-      end
-
-      def infer_status(attributes)
-        ProgrammeToIatiStatus.new.programme_status_to_iati_status(attributes[:programme_status])
       end
 
       def validate_from_codelist(code, entity, message)