+++ spec/features/staff/users_can_create_a_fund_level_activity_spec.rb
@@ -28,15 +28,6 @@ RSpec.feature "Users can create a fund level activity" do
       expect(activity.programme_status).to eq("spend_in_progress")
     end
 
-    scenario "the iati status gets set based on the default programme status value" do
-      identifier = "a-fund-with-default-iati-status-of-implementation"
-      visit activities_path
-      click_on(t("page_content.organisation.button.create_activity"))
-      fill_in_activity_form(delivery_partner_identifier: identifier, level: "fund")
-      activity = Activity.find_by(delivery_partner_identifier: identifier)
-      expect(activity.status).to eq("2")
-    end
-
     scenario "the activity form has some defaults" do
       activity = create(:fund_activity, organisation: user.organisation)
       activity_presenter = ActivityPresenter.new(activity)
@@ -49,20 +40,6 @@ RSpec.feature "Users can create a fund level activity" do
 
       visit activity_step_path(activity, :flow)
       expect(page.find("option[@selected = 'selected']").text).to eq activity_presenter.flow
-    end
-
-    scenario "the activity has the appropriate funding organisation defaults" do
-      identifier = "a-fund-has-a-funding-organisation"
-
-      visit activities_path
-      click_on(t("page_content.organisation.button.create_activity"))
-
-      fill_in_activity_form(delivery_partner_identifier: identifier, level: "fund")
-
-      activity = Activity.find_by(delivery_partner_identifier: identifier)
-      expect(activity.funding_organisation_name).to eq("HM Treasury")
-      expect(activity.funding_organisation_reference).to eq("GB-GOV-2")
-      expect(activity.funding_organisation_type).to eq("10")
     end
 
     scenario "the activity has the appropriate accountable organisation defaults" do